requirements.txt
faker>=0.7.0
gitpython
